+++ src/Engineering/Physics/Work.cpp
@@ -1,14 +1,14 @@
 #include "Work.h"
 
 // Set all the values to zero.
-Work::Work(float Mass){
+Work::Work(double Mass){
     mass = Mass; 
     TEnergy=0;
     listOfTypes = new LinkedList<EnergyValue>();
 }
 
 // compute all of the energy values based on inputs
-float Work::compute(){
+double Work::compute(){
     //Pg = mgh
     //GEnergy = mass*Gravity*height;
     //GSEnergy = 
@@ -20,7 +20,7 @@ float Work::compute(){
 }
 
 // returns the total watts from the system
-float Work::getTotalWatts(float time){
+double Work::getTotalWatts(double time){
     // prevent divide-by-zero
     if(time == 0) time = 0.00000000000001;
