+++ src/Environment/Planet/Planet_Atmosphere.cpp
@@ -1,7 +1,7 @@
 #include "Planet.h"
 
 // usesa sine wave to determine the density of the air according to the location.
-float Atmosphere::getDensity(float relative){
+double Atmosphere::getDensity(double relative){
      // D = (101325) / (287.05 * (15 + 273.15))
      // 
 }
