+++ src/Math/struct/vertex.h
@@ -13,61 +13,61 @@
 
 class vertex{
 public:
-	float x;
-        float y;
-        float z;
-	float ax; // Acceleration of x
-        float ay; // Acceleration of y
-        float az; // Acceleration of z
-        float mass; // for use with physics
-        float radius; // useful for sperical collision detection
-        float velocity; 
+	double x;
+        double y;
+        double z;
+	double ax; // Acceleration of x
+        double ay; // Acceleration of y
+        double az; // Acceleration of z
+        double mass; // for use with physics
+        double radius; // useful for sperical collision detection
+        double velocity;
 	long dimension; // Current Dimension (4, XYZt is default)
-	
-	vertex(float x,float y,float z);
+
+	vertex(double x,double y,double z);
         vertex();
         virtual ~vertex();
-	
+
 	vertex* operator =(vertex* b);
 	vertex* operator +=(vertex* b);
  	vertex* operator +(vertex* b);
  	vertex* operator -(vertex* b);
  	vertex* operator *(vertex* b);
- 	vertex* operator *(float b);
+ 	vertex* operator *(double b);
  	vertex* operator *(int b);
  	vertex* operator /(vertex* b);
- 	
+
         // convert to spherical coords
- 	float getRoh(vertex* body);
- 	float getPhi(vertex* body);
- 	float getTheta(vertex* body);
-        
-	float dot(vertex* v2); // Dot-product
+ 	double getRoh(vertex* body);
+ 	double getPhi(vertex* body);
+ 	double getTheta(vertex* body);
+
+	double dot(vertex* v2); // Dot-product
 	vertex* cross(vertex* v2); // Cross-product
 	vertex* unitVector(); // Directional vector
 	void add(vertex* v2); // add with vertex
 	void sub(vertex* v2);
-	void scale(float scalar);
+	void scale(double scalar);
 	void div(vertex* v2);
 	void mul(vertex* v2);
-	void set(float x, float y, float z);
-	float angle(vertex* v2); // angle from v2 based on 0 deg (X-direction)
-	float length(); // returns total distance between point and origin.
-	float length2(); // returns total distance between point and origin powered by 2
-	float distance(vertex* v2);
-	
+	void set(double x, double y, double z);
+	double angle(vertex* v2); // angle from v2 based on 0 deg (X-direction)
+	double length(); // returns total distance between point and origin.
+	double length2(); // returns total distance between point and origin powered by 2
+	double distance(vertex* v2);
+
 	// boolean functions to determine type.
 	bool isOrthagonal(vertex* v); // returns whether this is orthagonal.
-	
+
 	// Advanced Linear Algebra inequalities
 	// These are in-place, but have no apparent use.
 	bool cauchySchwarzInequality(vertex* v); // abs(dot(u,v)) <= length(u)*length(v)
 	bool triangleInequality(vertex* V); // length(u+v) <= length(u) + length(v)
 	bool pythagroreanInequality(vertex* v); // length(u+v)^2 = length(u)^2 + length(v)^2
-	
-	
+
+
 	//VectorSpace* addVectorSpace(vertex* b); // gets a vertex space from addition
-	//VectorSpace* scaleVectorSpace(float scalar); // gets a vertex space from scaling
+	//VectorSpace* scaleVectorSpace(double scalar); // gets a vertex space from scaling
 	//VectorSpace* getVectorSpace(); // gets a vertex space directly
 };
